requirements.txt
# Deps directes uniquement.
# On laisse pip résoudre les sous-dépendances compatibles avec Streamlit Cloud.

streamlit==1.56.0
python-dotenv==1.2.2
fastapi==0.116.1
uvicorn==0.35.0

dropbox==12.0.2
requests==2.33.1

openpyxl==3.1.5
pandas>=2.2,<3
tabulate==0.10.0

beautifulsoup4==4.14.3
PyMuPDF==1.27.2.2
pymupdf4llm==1.27.2.2
# Python 3.14.3 sur Streamlit Cloud (cf. logs du 2026-04-09).
python-docx==1.2.0
python-pptx==1.0.2

tiktoken==0.12.0
langchain-text-splitters==1.1.1
tqdm==4.67.3

openai==2.30.0
google-genai==1.69.0
